+++ server/apps/trak/tasks/__init__.py
@@ -1,2 +1,2 @@
 from .manifest_tasks import sync_site_manifests
-from .profile import sync_user_sites
+from .profile_tasks import sync_user_sites
